requirements.txt
openai==1.54.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
tiktoken==0.8.0
psycopg2-binary==2.9.9